Custom JSON Unmarashal for dates; Sort dates; Simple polling
Most of these are just proof-of-concepts for concepts I am reasonably sure will
be needed in the completed script.

Namely:
  1) Interpreting JSON data: This simplifies most other tasks working with the
     data
  2) Sorting date fields: Polling this data will require paging both over time
     and server responses.
  3) Polling: A worker will need to regularly poll the API for new data.

A deduper will also be required to ensure that only new alerts/events are
passed through the processing pipeline. This will likely need to track a list
of alert SIDs that have already been "seen". Keeping run state minimal will be
a primary focus for this component.