Add elementary map-based deduplication
More proof-of-concept work. Any polling will see the same event/alert multiple
times. This is a simple working deduplicator that tracks seen/observed Sids
using a map for fast lookups.

A better solution would be a LRU with time-based evictions that are configured
to always be a larger time window than the polling time window. This would
ensure that deduplication is guaranteed, while also ensuring that the cache
growth is bounded.

It will probably also be a good idea to provide an option that allows the cache
to be pre-populated before it is used for deduplication. This would further cut
down on the chance of emitting duplicate events when stopping and restarting
the monitor, but also does not introduce state that must be tracked between
invocations.